Move unit tests in their own folder
