+++ packages/tests/jest.config.js
@@ -9,7 +9,7 @@ module.exports = {
     preset: 'ts-jest/presets/js-with-ts',
     // Without it, vite import is silently crashing the process with code SIGHUP 129
     resetModules: true,
-    roots: ['./src/'],
+    roots: ['./src/unit/'],
     setupFilesAfterEnv: ['<rootDir>/src/_jest/setupAfterEnv.ts'],
     testEnvironment: 'node',
     testMatch: ['**/*.test.*'],
