+++ packages/tests/src/unit/plugins/error-tracking/index.test.ts
