+++ packages/tests/src/unit/plugins/error-tracking/testHelpers.ts
