+++ packages/tests/src/unit/plugins/error-tracking/validate.test.ts
