+++ packages/tests/src/unit/plugins/telemetry/.eslintrc.js
