+++ packages/tests/src/unit/plugins/telemetry/testHelpers.ts
