+++ packages/tests/src/unit/tools/src/commands/create-plugin/hooks.test.ts
