+++ packages/tests/src/unit/tools/src/commands/create-plugin/index.test.ts
